# lib/api-zod/src/generated/types/url.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API contract for the Distributed URL Shortener
 * OpenAPI spec version: 0.1.0
 */

export interface Url {
  shortCode: string;
  shortUrl: string;
  originalUrl: string;
  createdAt: Date;
  /** @nullable */
  expiresAt: Date | null;
  active: boolean;
  clickCount: number;
  updatedAt: Date;
}