# lib/api-zod/src/generated/types/urlStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API contract for the Distributed URL Shortener
 * OpenAPI spec version: 0.1.0
 */

export interface UrlStats {
  shortCode: string;
  clickCount: number;
  active: boolean;
  /** @nullable */
  expiresAt: Date | null;
  lastUpdatedAt: Date;
}